pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py>=2.5.2",
    "google-api-python-client>=2.172.0",
    "python-dotenv>=1.1.0",
]
